Move nickname to a property of socket
Instead of making sockets[] an array of objects containing socket and
name it now just contains socket (as it was originally) with nickname as
a property of socket. This redundifies the findExistingName() function
because we can use the existing indexOf() and just use socket.nickname
instead of some long array stuff. This makes for less code! ;)